Move init functions from routers/install to routers/init (#230)
* move init functions from routers/install to routers/init

* copyright typo
